% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure-model.R
\name{configure_model}
\alias{configure_model}
\title{Configure (but do not fit) a model and associated variables}
\usage{
configure_model(self)
}
\arguments{
\item{self}{Metaflow state variable}
}
\description{
Configure (but do not fit) a model and associated variables
}
