% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-evaluation.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluate a Yelp review model}
\usage{
evaluate_model(model, test)
}
\arguments{
\item{model}{Fitted \code{parsnip} or \code{workflows} model}

\item{test}{Test data containing "text" and "stars" columns}
}
\value{
tibble
}
\description{
Evaluate a Yelp review model
}
