% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-flow.R
\name{generate_flow}
\alias{generate_flow}
\title{Generate model training flow}
\usage{
generate_flow()
}
\value{
Metaflow flow object
}
\description{
This function requires that an AWS account ID is configured as an
environment variable "AWS_ACCOUNT_ID". This is used to construct the
location of the image used to run the flow on AWS Batch.
}
