% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data.R
\name{generate_text_processing_recipe}
\alias{generate_text_processing_recipe}
\title{Generate a recipe for processing text data into a document-term matrix}
\usage{
generate_text_processing_recipe(formula, train_data, min_times = 10)
}
\arguments{
\item{formula}{A model formula.}

\item{train_data}{A data frame or tibble of the \emph{template} data set.}

\item{min_times}{An integer. Minimum number of times a word can appear before
getting removed. Defaults to 10.}
}
\value{
\code{recipes::recipe} object
}
\description{
Adapted from work by Emil Hvitfeldt
https://www.hvitfeldt.me/blog/text-classification-with-tidymodels/
}
