% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-data.R
\name{prepare_data}
\alias{prepare_data}
\title{Step function for preparing and splitting data}
\usage{
prepare_data(self)
}
\arguments{
\item{self}{Metaflow state variable}
}
\description{
Step function for preparing and splitting data
}
