% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameter-tuning.R
\name{select_best_hyperparameters}
\alias{select_best_hyperparameters}
\title{Select the best hyperparameter set according to a given metric}
\usage{
select_best_hyperparameters(hyperparameter_results, metric = "rmse")
}
\arguments{
\item{hyperparameter_results}{A tibble of hyperparameter results collected
with \code{tune::collect_metrics}}

\item{metric}{A metric to minimise. Defaults to "rmse".}
}
\value{
named list
}
\description{
Metaflow saves artefacts as pickle files, which prevents us from using nested
tibbles. Unfortunately, \code{tune} relies on nested tibbles to store its results.
We instead use \code{collect_metrics} to store simple tibbles that are compatible
with pickle. Unfortunately, this means we cannot rely on \code{tune}'s functions
for extracting optimal hyperparameter sets and must write our own.
}
