% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-final-model.R
\name{train_final_model}
\alias{train_final_model}
\title{Compare hyperparameter performance and train a final model}
\usage{
train_final_model(self, inputs)
}
\arguments{
\item{self}{Metaflow state variable}

\item{inputs}{Inputs from the previous split Metaflow steps}
}
\description{
Compare hyperparameter performance and train a final model
}
\details{
This step is join in the flow, and so must first reduce the various
hyperparameter evaluations produced in the previous steps and merge the
other variables.

It appears as though Metaflow struggles with nested tibbles, likely due to
a restriction with reticulate/Python. This prevents us from using the
tuning results from \code{tune} directly, so we have to use a custom
\code{select_best_hyperparameters} function.
}
