% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune-hyperparameters.R
\name{tune_hyperparameters}
\alias{tune_hyperparameters}
\title{Tune a subset of hyperparameters}
\usage{
tune_hyperparameters(self)
}
\arguments{
\item{self}{Metaflow state variable}
}
\description{
Tune a subset of hyperparameters
}
\details{
This step is split in the flow so that separate instances can evaluate
different combinations of hyperparameters indepentently of one another.

It appears as though Metaflow struggles with nested tibbles, likely due to
a restriction with reticulate/Python. This prevents us from using the
tuning results from \code{tune} directly, so we have to \code{collect_metrics}.
}
